requirements.txt
flask
gevent-websocket
websocket-client
